refactor: 예외 및 설정 클래스 전역 패키지로 리팩토링
