+++ src/main/java/roomescape/global/config/JwtConfiguration.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.global.config;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
