+++ src/main/java/roomescape/global/config/TimeConfiguration.java
@@ -1,7 +1,6 @@
-package roomescape.config;
+package roomescape.global.config;
 
 import java.time.Clock;
-
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
