+++ src/main/java/roomescape/global/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class ForbiddenException extends RuntimeException {
