+++ src/main/java/roomescape/global/exception/PaymentClientException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class PaymentClientException extends RuntimeException {
