+++ src/main/java/roomescape/global/exception/ReservationException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class ReservationException extends RuntimeException {
