+++ src/main/java/roomescape/global/exception/RoomescapeException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class RoomescapeException extends RuntimeException {
