+++ src/main/java/roomescape/global/exception/TokenCreationException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.global.exception;
 
 public class TokenCreationException extends RuntimeException {
