+++ src/main/java/roomescape/reservation/config/TossClientErrorHandler.java
@@ -10,8 +10,8 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.ResponseErrorHandler;
-import roomescape.exception.PaymentClientException;
-import roomescape.exception.RoomescapeException;
+import roomescape.global.exception.PaymentClientException;
+import roomescape.global.exception.RoomescapeException;
 
 @Slf4j
 @Component
