+++ src/test/java/roomescape/reservation/config/TossClientErrorHandlerTest.java
@@ -18,8 +18,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.client.ClientHttpResponse;
-import roomescape.exception.PaymentClientException;
-import roomescape.exception.RoomescapeException;
+import roomescape.global.exception.PaymentClientException;
+import roomescape.global.exception.RoomescapeException;
 
 @ExtendWith(MockitoExtension.class)
 class TossClientErrorHandlerTest {
