+++ src/test/java/roomescape/reservation/config/TossServerErrorHandlerTest.java
@@ -15,7 +15,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.client.ClientHttpResponse;
-import roomescape.exception.RoomescapeException;
+import roomescape.global.exception.RoomescapeException;
 
 @ExtendWith(MockitoExtension.class)
 class TossServerErrorHandlerTest {
