+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -20,9 +20,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.auth.dto.LoginMember;
-import roomescape.exception.NotFoundException;
-import roomescape.exception.PaymentClientException;
-import roomescape.exception.ReservationException;
+import roomescape.global.exception.NotFoundException;
+import roomescape.global.exception.PaymentClientException;
+import roomescape.global.exception.ReservationException;
 import roomescape.member.domain.Member;
 import roomescape.member.repository.MemberRepository;
 import roomescape.reservation.BaseTest;
